+++ src/main/java/br/edu/ifpb/dac/igorsobral/projetojpa/dao/InstituicaoDAORepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.ifpb.dac.igorsobral.projetojpa.entity.Instituicao;
 
 
-public interface InstituicaoDAORepository extends JpaRepository<Instituicao, Integer>{
+public interface InstituicaoDAORepository extends JpaRepository<Instituicao, Long>{
 
 }
